[DoctrineBridge] Extracted the common type and made the choice list generic
