+++ src/Symfony/Bridge/Doctrine/Form/Type/EntityType.php
@@ -11,91 +11,28 @@
 
 namespace Symfony\Bridge\Doctrine\Form\Type;
 
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Common\Persistence\ObjectManager;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Bridge\Doctrine\Form\ChoiceList\EntityChoiceList;
+use Symfony\Bridge\Doctrine\Form\ChoiceList\EntityLoaderInterface;
 use Symfony\Bridge\Doctrine\Form\ChoiceList\ORMQueryBuilderLoader;
-use Symfony\Bridge\Doctrine\Form\EventListener\MergeCollectionListener;
-use Symfony\Bridge\Doctrine\Form\DataTransformer\EntitiesToArrayTransformer;
-use Symfony\Bridge\Doctrine\Form\DataTransformer\EntityToIdTransformer;
-use Symfony\Component\Form\AbstractType;
 
-class EntityType extends AbstractType
+class EntityType extends DoctrineType
 {
-    /**
-     * @var ManagerRegistry
-     */
-    protected $registry;
-
-    public function __construct(ManagerRegistry $registry)
-    {
-        $this->registry = $registry;
-    }
-
-    public function buildForm(FormBuilder $builder, array $options)
-    {
-        if ($options['multiple']) {
-            $builder
-                ->addEventSubscriber(new MergeCollectionListener())
-                ->prependClientTransformer(new EntitiesToArrayTransformer($options['choice_list']))
-            ;
-        } else {
-            $builder->prependClientTransformer(new EntityToIdTransformer($options['choice_list']));
-        }
-    }
-
-    public function getDefaultOptions(array $options)
-    {
-        $defaultOptions = array(
-            'em'                => null,
-            'class'             => null,
-            'property'          => null,
-            'query_builder'     => null,
-            'loader'            => null,
-            'choices'           => null,
-            'group_by'          => null,
-        );
-
-        $options = array_replace($defaultOptions, $options);
-
-        if (!isset($options['choice_list'])) {
-            $manager = $this->registry->getManager($options['em']);
-            if (isset($options['query_builder'])) {
-                $options['loader'] = $this->getLoader($manager, $options);
-            }
-
-            $defaultOptions['choice_list'] = new EntityChoiceList(
-                $manager,
-                $options['class'],
-                $options['property'],
-                $options['loader'],
-                $options['choices'],
-                $options['group_by']
-            );
-        }
-
-        return $defaultOptions;
-    }
-
     /**
      * Return the default loader object.
      *
-     * @param ObjectManager
+     * @param ObjectManager $manager
      * @param array $options
      * @return ORMQueryBuilderLoader
      */
-    protected function getLoader($manager, $options)
+    protected function getLoader(ObjectManager $manager, array $options)
     {
         return new ORMQueryBuilderLoader(
             $options['query_builder'],
             $manager,
             $options['class']
         );
-    }
-
-    public function getParent(array $options)
-    {
-        return 'choice';
     }
 
     public function getName()
